Add Enums (FileType, Command)
