+++ core/src/main/java/core/FileInfo.java
@@ -1,30 +1,15 @@
 package core;
 
 import java.io.*;
-import java.nio.ByteBuffer;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
 public class FileInfo implements Serializable {
 
-    public enum FileType {
-        FILE("File"), DIRECTORY("Directory");
-
-        private String name;
-
-        public String getName() {
-            return name;
-        }
-
-        FileType(String name) {
-            this.name = name;
-        }
-    }
-
     private String filename;
     private long size;
     private FileType type;
-    private String path;
+    private final String path;
 
     public FileInfo(Path path){
         try {
@@ -33,7 +18,7 @@ public class FileInfo implements Serializable {
             this.size = Files.size(path);
             this.type = Files.isDirectory(path) ? FileType.DIRECTORY : FileType.FILE;
             if (this.type == FileType.DIRECTORY) {
-                this.size = -1L;
+                this.size = 0;
             }
         } catch (IOException e) {
             throw new RuntimeException("Can't get a info from file");
