+++ core/src/main/java/core/FileType.java
@@ -1,0 +1,16 @@
+package core;
+
+public enum FileType {
+    FILE("File"), DIRECTORY("DIR");
+
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    FileType(String name) {
+        this.name = name;
+    }
+}
+
