+++ server/src/main/java/ServerMessageHandler.java
@@ -1,21 +1,17 @@
 
+import core.Command;
 import core.Message;
-import core.Receiver;
 import core.Sender;
 
 import java.io.IOException;
 import java.nio.channels.SocketChannel;
-
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
 public class ServerMessageHandler {
 
-    private Message message;
-    private SocketChannel channel;
-    private Receiver receiver;
-    private Sender sender;
-    private Path path;
+    private final Message message;
+    private final SocketChannel channel;
 
     public ServerMessageHandler(Message message, SocketChannel channel) throws IOException {
         this.message = message;
@@ -25,10 +21,10 @@ public class ServerMessageHandler {
 
 
     private void handle() throws IOException {
-        if (message.getCmd() == Message.Command.GET_LIST){
+        if (message.getCmd() == Command.GET_LIST){
             String stringPath = message.getFileInfo().getStringPath();
-            path = Paths.get(stringPath);
-            sender = new Sender(this.channel, path);
+            Path path = Paths.get(stringPath);
+            Sender sender = new Sender(this.channel, path);
             try {
                 sender.sendFilesList();
                 channel.close();
